requirements.txt
openrgb-python
requests
jsonschema
python-dotenv
numpy
colour-science
logger_tt
